feat(sheet): extract pure selection core
Move clampSelection, clampRange, getPrimaryCell, isCellInSelection,
getSelectionAddress, getColumnLabel into sheet/core/selection.ts and
extract nextSelectionForKey (arrow/Tab/Enter nav math). SheetView imports
them and delegates grid-key navigation to the pure function.

Spec fix locked by tests: Shift+Tab at 0,0 now stays put rather than
jumping to the last column of row 0.
